clinical_trial_pipeline_bit_surgeons/requirements.txt
numpy==1.24.4
pandas==2.0.3
scikit-learn==1.3.2
sentence-transformers==3.1.1
umap-learn==0.5.7
hdbscan==0.8.40
torch==2.3.1
transformers==4.45.2
faiss-gpu==1.7.2
matplotlib==3.7.5
nltk==3.9.1
datasets==2.21.0
